Apollo-4222 Refactoring the spark-solr codebase
Clean up the code around SolrRDD and SolrRelation to make them simpler and thus, easier to port to Scala.

The core changes done in this PR:

  Move all the static methods to utility classes
  Untie dependency of schema from core SolrRDD class
  Got rid of queryDeep method in SolrRDD
  Added tests for RDD and SQL queries
  Removed SchemaPreservingRDD. We might add this later.
  Add coverage profile to maven

The core logic did not change much in this branch. The refactoring is to make it simpler to work with the codebase.
