+++ src/main/java/com/lucidworks/spark/example/query/KMeansAnomaly.java
@@ -1,8 +1,9 @@
 package com.lucidworks.spark.example.query;
 
-import com.lucidworks.spark.SolrRDD;
-import com.lucidworks.spark.SolrRelation;
-import com.lucidworks.spark.SolrSupport;
+import com.lucidworks.spark.util.SolrQuerySupport;
+import com.lucidworks.spark.util.SolrQuerySupport.PivotField;
+import com.lucidworks.spark.rdd.SolrRDD;
+import com.lucidworks.spark.util.SolrSupport;
 import com.lucidworks.spark.SparkApp;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
@@ -11,6 +12,9 @@ import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.api.java.function.Function;
+import org.apache.spark.mllib.clustering.KMeans;
+import org.apache.spark.mllib.clustering.KMeansModel;
+import org.apache.spark.mllib.linalg.Vector;
 import org.apache.spark.mllib.linalg.Vectors;
 import org.apache.spark.sql.DataFrame;
 import org.apache.spark.sql.Row;
@@ -20,15 +24,11 @@ import org.apache.spark.sql.api.java.UDF1;
 import org.apache.spark.sql.hive.HiveContext;
 import org.apache.spark.sql.types.DataTypes;
 
-import org.apache.spark.mllib.clustering.KMeans;
-import org.apache.spark.mllib.clustering.KMeansModel;
-import org.apache.spark.mllib.linalg.Vector;
-
-import com.lucidworks.spark.SolrRDD.PivotField;
-
 import java.sql.Timestamp;
 import java.util.HashMap;
 import java.util.Map;
+
+import static com.lucidworks.spark.util.ConfigurationConstants.*;
 
 /**
  * Use K-means to do basic anomaly detection by examining user sessions
@@ -75,9 +75,9 @@ public class KMeansAnomaly implements SparkApp.RDDProcessor {
     options.put("zkhost", zkHost);
     options.put("collection", collection);
     options.put("query", queryStr);
-    options.put(SolrRDD.SOLR_SPLIT_FIELD_PARAM, "_version_");
-    options.put(SolrRDD.SOLR_SPLITS_PER_SHARD_PARAM, "4");
-    options.put(SolrRDD.SOLR_FIELD_LIST_PARAM, "id,_version_,"+UID_FIELD+","+TS_FIELD+",bytes_s,response_s,verb_s");
+    options.put(SOLR_SPLIT_FIELD_PARAM, "_version_");
+    options.put(SOLR_SPLITS_PER_SHARD_PARAM, "4");
+    options.put(SOLR_FIELD_LIST_PARAM, "id,_version_,"+UID_FIELD+","+TS_FIELD+",bytes_s,response_s,verb_s");
 
     // Use the Solr DataSource to load rows from a Solr collection using a query
     // highlights include:
@@ -98,7 +98,8 @@ public class KMeansAnomaly implements SparkApp.RDDProcessor {
     };
 
     // this "view" has the verb_s and response_s fields expanded into aggregatable
-    DataFrame solrDataWithPivots = (new SolrRDD(zkHost, collection)).withPivotFields(logEvents, pivotFields);
+    SolrRDD solrRDD = new SolrRDD(zkHost, collection, jsc.sc());
+    DataFrame solrDataWithPivots = SolrQuerySupport.withPivotFields(logEvents, pivotFields, solrRDD, false);
     // register this DataFrame so we can execute a SQL query against it for doing sessionization using lag window func
     solrDataWithPivots.registerTempTable("logs");
