+++ src/main/java/com/lucidworks/spark/query/PagedResultsIterator.java
@@ -1,8 +1,8 @@
 package com.lucidworks.spark.query;
 
-import com.lucidworks.spark.SolrRDD;
-import org.apache.solr.client.solrj.SolrQuery;
+import com.lucidworks.spark.util.SolrQuerySupport;
 import org.apache.solr.client.solrj.SolrClient;
+import org.apache.solr.client.solrj.SolrQuery;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.impl.CloudSolrClient;
 import org.apache.solr.client.solrj.response.QueryResponse;
@@ -70,7 +70,7 @@ public abstract class PagedResultsIterator<T> implements Iterator<T>, Iterable<T
 
   protected List<T> fetchNextPage() throws SolrServerException {
     int start = (cursorMark != null) ? 0 : getStartForNextPage();
-    QueryResponse resp = SolrRDD.querySolr(solrServer, solrQuery, start, cursorMark);
+    QueryResponse resp = SolrQuerySupport.querySolr(solrServer, solrQuery, start, cursorMark);
     if (cursorMark != null)
       cursorMark = resp.getNextCursorMark();
