+++ src/main/java/com/lucidworks/spark/query/ShardSplit.java
@@ -1,6 +1,7 @@
 package com.lucidworks.spark.query;
 
 import org.apache.solr.client.solrj.SolrQuery;
+
 import java.io.Serializable;
 
 /**
