--- src/main/java/com/lucidworks/spark/streaming/MessageHandler.java
@@ -1,6 +1,0 @@
-package com.lucidworks.spark.streaming;
-
-public interface MessageHandler {
-    void handleMessage(String msg);
-    void onEndOfStream();
-}
