--- src/main/java/com/lucidworks/spark/streaming/MessageStream.java
@@ -1,10 +1,0 @@
-package com.lucidworks.spark.streaming;
-
-import java.io.Serializable;
-import java.net.URI;
-
-public interface MessageStream extends Serializable {
-    URI getURI();
-    void receive(MessageHandler handler);
-    void stopReceiving();
-}
