+++ src/main/java/com/lucidworks/spark/util/ConfigurationConstants.java
@@ -1,0 +1,15 @@
+package com.lucidworks.spark.util;
+
+public interface ConfigurationConstants {
+
+  public static final String CONFIG_PREFIX = "solr.";
+
+  public static final String SOLR_ZK_HOST_PARAM = "zkhost";
+  public static final String SOLR_COLLECTION_PARAM = "collection";
+  public static final String SOLR_QUERY_PARAM = "query";
+  public static final String SOLR_FIELD_LIST_PARAM = "query.fields";
+  public static final String SOLR_ROWS_PARAM = "query.rows";
+  public static final String SOLR_SPLIT_FIELD_PARAM = "query.split.field";
+  public static final String SOLR_SPLITS_PER_SHARD_PARAM = "query.splits.per.shard";
+  public static final String ESCAPE_FIELDNAMES = "escape.fieldnames";
+}
