+++ src/main/java/com/lucidworks/spark/util/QueryConstants.java
@@ -1,0 +1,8 @@
+package com.lucidworks.spark.util;
+
+public class QueryConstants {
+  public static final String DEFAULT_REQUIRED_FIELD = "id";
+  public static final int DEFAULT_PAGE_SIZE = 1000;
+  public static final String DEFAULT_QUERY = "*:*";
+  public static final int DEFAULT_SPLITS_PER_SHARD = 20;
+}
