+++ src/test/java/com/lucidworks/spark/RDDProcessorTestBase.java
@@ -1,5 +1,7 @@
 package com.lucidworks.spark;
 
+import com.lucidworks.spark.rdd.SolrRDD;
+import com.lucidworks.spark.util.SolrSupport;
 import org.apache.solr.common.SolrDocument;
 import org.apache.solr.common.SolrInputDocument;
 import org.apache.spark.SparkConf;
@@ -67,8 +69,8 @@ public class RDDProcessorTestBase extends TestSolrCloudClusterSupport implements
       int numDocsIndexed = indexDocs(zkHost, collection, inputDocs);
       Thread.sleep(1000L);
       // verify docs got indexed ... relies on soft auto-commits firing frequently
-      SolrRDD solrRDD = new SolrRDD(zkHost, collection);
-      JavaRDD<SolrDocument> resultsRDD = solrRDD.query(jsc.sc(), "*:*");
+      SolrRDD solrRDD = new SolrRDD(zkHost, collection, jsc.sc());
+      JavaRDD<SolrDocument> resultsRDD = solrRDD.query("*:*");
       long numFound = resultsRDD.count();
       assertTrue("expected " + numDocsIndexed + " docs in query results from " + collection + ", but got " + numFound,
           numFound == (long) numDocsIndexed);
