+++ src/test/java/com/lucidworks/spark/SolrRelationTest.java
@@ -19,11 +19,11 @@ import org.apache.spark.mllib.linalg.Vectors;
 import org.apache.commons.io.FileUtils;
 import solr.DefaultSource;
 
-import java.io.File;
+import java.io.*;
 import java.util.*;
 
 import static org.junit.Assert.*;
-
+import static com.lucidworks.spark.util.ConfigurationConstants.*;
 
 /**
  * Tests for the SolrRelation implementation.
@@ -77,13 +77,13 @@ public class SolrRelationTest extends RDDProcessorTestBase {
     deleteCollection(testCollection);
   }
   
-  protected String array2cdl(String[] arr) {
+  protected static String array2cdl(String[] arr) {
     // this is really horrible
     String str = Arrays.asList(arr).toString();
     return str.substring(1, str.length() - 1).replaceAll(" ","");
   }
   
-  protected Row[] validateDataFrameStoreLoad(SQLContext sqlContext, String testCollection, DataFrame sourceData) throws Exception {
+  protected static Row[] validateDataFrameStoreLoad(SQLContext sqlContext, String testCollection, DataFrame sourceData) throws Exception {
     sourceData = sourceData.sort("id");
     sourceData.printSchema();
     Row[] testData = sourceData.collect();
@@ -91,8 +91,8 @@ public class SolrRelationTest extends RDDProcessorTestBase {
 
     String zkHost = cluster.getZkServer().getZkAddress();
     Map<String, String> options = new HashMap<String, String>();
-    options.put(SolrRDD.SOLR_ZK_HOST_PARAM, zkHost);
-    options.put(SolrRDD.SOLR_COLLECTION_PARAM, testCollection);
+    options.put(SOLR_ZK_HOST_PARAM, zkHost);
+    options.put(SOLR_COLLECTION_PARAM, testCollection);
     sourceData.write().format(DefaultSource.SOLR_FORMAT).options(options).mode(SaveMode.Overwrite).save();
     Thread.sleep(1000);
 
@@ -102,7 +102,7 @@ public class SolrRelationTest extends RDDProcessorTestBase {
     dumpSolrCollection(testCollection, q);
 
     // now read the data back from Solr and validate that it was saved correctly and that all data type handling is correct
-    options.put(SolrRDD.SOLR_FIELD_LIST_PARAM, array2cdl(cols));
+    options.put(SOLR_FIELD_LIST_PARAM, array2cdl(cols));
     DataFrame fromSolr = sqlContext.read().format(DefaultSource.SOLR_FORMAT).options(options).load();
     fromSolr = fromSolr.sort("id");
     fromSolr.printSchema();
@@ -149,11 +149,11 @@ public class SolrRelationTest extends RDDProcessorTestBase {
     buildCollection(zkHost, testCollection, testData, 2);
 
     Map<String, String> options = new HashMap<String, String>();
-    options.put(SolrRDD.SOLR_ZK_HOST_PARAM, zkHost);
-    options.put(SolrRDD.SOLR_COLLECTION_PARAM, testCollection);
+    options.put(SOLR_ZK_HOST_PARAM, zkHost);
+    options.put(SOLR_COLLECTION_PARAM, testCollection);
 
     DataFrame df = sqlContext.read().format("solr").options(options).load();
-
+    df.show();
     validateSchema(df);
     //df.show();
 
@@ -212,8 +212,8 @@ public class SolrRelationTest extends RDDProcessorTestBase {
     createCollection("testFilterSupport2", numShards, replicationFactor, 2, confName, confDir);
 
     options = new HashMap<String, String>();
-    options.put(SolrRDD.SOLR_ZK_HOST_PARAM, zkHost);
-    options.put(SolrRDD.SOLR_COLLECTION_PARAM, "testFilterSupport2");
+    options.put(SOLR_ZK_HOST_PARAM, zkHost);
+    options.put(SOLR_COLLECTION_PARAM, "testFilterSupport2");
 
     df.write().format("solr").options(options).mode(SaveMode.Overwrite).save();
     Thread.sleep(1000);
@@ -223,145 +223,6 @@ public class SolrRelationTest extends RDDProcessorTestBase {
 
     deleteCollection(testCollection);
     deleteCollection("testFilterSupport2");
-  }
-
-  //@Ignore
-  @Test
-  public void testNestedDataFrames() throws Exception {
-    SQLContext sqlContext = new SQLContext(jsc);
-    String confName = "testConfig";
-    File confDir = new File("src/test/resources/conf");
-    int numShards = 2;
-    int replicationFactor = 1;
-    deleteCollection("testNested");
-    createCollection("testNested", numShards, replicationFactor, 2, confName, confDir);
-    List<StructField> fields = new ArrayList<StructField>();
-    List<StructField> fields1 = new ArrayList<StructField>();
-    List<StructField> fields2 = new ArrayList<StructField>();
-    fields.add(DataTypes.createStructField("id", DataTypes.StringType, true));
-    fields.add(DataTypes.createStructField("testing_s", DataTypes.StringType, true));
-    fields1.add(DataTypes.createStructField("test1_s", DataTypes.StringType, true));
-    fields1.add(DataTypes.createStructField("test2_s", DataTypes.StringType, true));
-    fields2.add(DataTypes.createStructField("test11_s", DataTypes.StringType, true));
-    fields2.add(DataTypes.createStructField("test12_s", DataTypes.StringType, true));
-    fields2.add(DataTypes.createStructField("test13_s", DataTypes.StringType, true));
-    fields1.add(DataTypes.createStructField("testtype_s", DataTypes.createStructType(fields2), true));
-    fields.add(DataTypes.createStructField("test_s", DataTypes.createStructType(fields1), true));
-    StructType schema = DataTypes.createStructType(fields);
-    Row dm = RowFactory.create("7", "test", RowFactory.create("test1", "test2", RowFactory.create("test11", "test12", "test13")));
-    List<Row> list = new ArrayList<Row>();
-    list.add(dm);
-    JavaRDD<Row> rdd = jsc.parallelize(list);
-    DataFrame df = sqlContext.createDataFrame(rdd, schema);
-    HashMap<String, String> options = new HashMap<String, String>();
-    String zkHost = cluster.getZkServer().getZkAddress();
-    options = new HashMap<String, String>();
-    options.put(SolrRDD.SOLR_ZK_HOST_PARAM, zkHost);
-    options.put(SolrRDD.SOLR_COLLECTION_PARAM, "testNested");
-    options.put("preserveschema", "Y");
-    df.write().format("solr").options(options).mode(SaveMode.Overwrite).save();
-    Thread.sleep(1000);
-    DataFrame df2 = sqlContext.read().format("solr").options(options).load();
-    df2 = sqlContext.createDataFrame(df2.javaRDD(),df2.schema());
-    df.show();
-    df2.show();
-    df2.registerTempTable("DFTEST");
-    sqlContext.sql("SELECT test_s.testtype_s FROM DFTEST").show();
-    deleteCollection("testNested");
-  }
-
-  public void createMLModelLRParquet() throws Exception {
-    File lRModel = new File("LRParquet").getAbsoluteFile();
-    if (lRModel.exists()) FileUtils.forceDelete(lRModel);
-    List<LabeledPoint> list = new ArrayList<LabeledPoint>();
-    LabeledPoint zero = new LabeledPoint(0.0, Vectors.dense(1.0, 0.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0));
-    LabeledPoint one = new LabeledPoint(1.0, Vectors.dense(8.0,7.0,6.0,4.0,5.0,6.0,1.0,2.0,3.0));
-    list.add(zero);
-    list.add(one);
-    JavaRDD<LabeledPoint> data = jsc.parallelize(list);
-    final LogisticRegressionModel model = new LogisticRegressionWithLBFGS()
-              .setNumClasses(2)
-              .run(data.rdd());
-    model.save(jsc.sc(), "LRParquet");
-  }
-
-  public void createMLModelNBParquet() throws Exception {
-    File nBModel = new File("NBParquet").getAbsoluteFile();
-    if (nBModel.exists()) FileUtils.forceDelete(nBModel);
-    List<LabeledPoint> list = new ArrayList<LabeledPoint>();
-    LabeledPoint zero = new LabeledPoint(0.0, Vectors.dense(1.0, 0.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0));
-    LabeledPoint one = new LabeledPoint(1.0, Vectors.dense(8.0,7.0,6.0,4.0,5.0,6.0,1.0,2.0,3.0));
-    list.add(zero);
-    list.add(one);
-    JavaRDD<LabeledPoint> data = jsc.parallelize(list);
-    final NaiveBayesModel model = NaiveBayes.train(data.rdd(), 1.0);
-    model.save(jsc.sc(), "NBParquet");
-  }
-
-  //@Ignore
-  @Test
-  public void loadLRParquetIntoSolr() throws Exception {
-    createMLModelLRParquet();
-    SQLContext sqlContext = new SQLContext(jsc);
-    String confName = "testConfig";
-    File confDir = new File("src/test/resources/conf");
-    int numShards = 2;
-    int replicationFactor = 1;
-    deleteCollection("TestLR");
-    Thread.sleep(1000);
-    createCollection("TestLR", numShards, replicationFactor, 2, confName, confDir);
-    String zkHost = cluster.getZkServer().getZkAddress();
-    DataFrame dfLR = sqlContext.load("LRParquet/data/");
-    HashMap<String, String> options = new HashMap<String, String>();
-    options = new HashMap<String, String>();
-    options.put(SolrRDD.SOLR_ZK_HOST_PARAM, zkHost);
-    options.put(SolrRDD.SOLR_COLLECTION_PARAM, "TestLR");
-    options.put("preserveschema", "Y");
-    dfLR.write().format("solr").options(options).mode(SaveMode.Overwrite).save();
-    dfLR.show();
-    dfLR.printSchema();
-    Thread.sleep(5000);
-    DataFrame dfLR2 = sqlContext.read().format("solr").options(options).load();
-    dfLR2.show();
-    dfLR2.printSchema();
-    assertCount(dfLR.count(), dfLR.intersect(dfLR2).count(), "compare dataframe count");
-    deleteCollection("TestLR");
-    Thread.sleep(1000);
-    File lRModel = new File("LRParquet").getAbsoluteFile();
-    FileUtils.forceDelete(lRModel);
-  }
-
-  //@Ignore
-  @Test
-  public void loadNBParquetIntoSolr() throws Exception {
-    createMLModelNBParquet();
-    SQLContext sqlContext = new SQLContext(jsc);
-    String confName = "testConfig";
-    File confDir = new File("src/test/resources/conf");
-    int numShards = 2;
-    int replicationFactor = 1;
-    deleteCollection("TestNB");
-    Thread.sleep(1000);
-    createCollection("TestNB", numShards, replicationFactor, 2, confName, confDir);
-    String zkHost = cluster.getZkServer().getZkAddress();
-    DataFrame dfNB = sqlContext.load("NBParquet/data/");
-    HashMap<String, String> options = new HashMap<String, String>();
-    options = new HashMap<String, String>();
-    options.put(SolrRDD.SOLR_ZK_HOST_PARAM, zkHost);
-    options.put("preserveschema", "Y");
-    options.put(SolrRDD.SOLR_COLLECTION_PARAM, "TestNB");
-    dfNB.write().format("solr").options(options).mode(SaveMode.Overwrite).save();
-    dfNB.show();
-    Thread.sleep(5000);
-    DataFrame dfNB2 = sqlContext.read().format("solr").options(options).load();
-    dfNB2.show();
-    dfNB.printSchema();
-    dfNB2.printSchema();
-    assertCount(dfNB.count(), dfNB.intersect(dfNB2).count(), "compare dataframe count");
-    deleteCollection("TestNB");
-    Thread.sleep(1000);
-    File nBModel = new File("NBParquet").getAbsoluteFile();
-    FileUtils.forceDelete(nBModel);
   }
 
   protected void assertCount(long expected, long actual, String expr) {
