+++ src/test/java/com/lucidworks/spark/example/query/BuildQueryTest.java
@@ -1,6 +1,7 @@
 package com.lucidworks.spark.example.query;
 
-import com.lucidworks.spark.SolrRDD;
+import com.lucidworks.spark.rdd.SolrRDD;
+import com.lucidworks.spark.util.SolrQuerySupport;
 import org.apache.solr.client.solrj.SolrQuery;
 import org.junit.Test;
 
@@ -11,6 +12,7 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static com.lucidworks.spark.util.QueryConstants.*;
 
 public class BuildQueryTest {
 
@@ -18,20 +20,20 @@ public class BuildQueryTest {
   public void testQueryBuilder() {
     SolrQuery q = null;
 
-    q = SolrRDD.toQuery(null);
+    q = SolrQuerySupport.toQuery(null);
     assertEquals("*:*", q.getQuery());
-    assertEquals(new Integer(SolrRDD.DEFAULT_PAGE_SIZE), q.getRows());
+    assertEquals(new Integer(DEFAULT_PAGE_SIZE), q.getRows());
 
-    q = SolrRDD.toQuery("q=*:*");
+    q = SolrQuerySupport.toQuery("q=*:*") ;
     assertEquals("*:*", q.getQuery());
-    assertEquals(new Integer(SolrRDD.DEFAULT_PAGE_SIZE), q.getRows());
+    assertEquals(new Integer(DEFAULT_PAGE_SIZE), q.getRows());
 
     String qs = "text:hello";
     String fq = "price:[100 TO *]";
     String sort = "id";
-    q = SolrRDD.toQuery("q="+encode(qs)+"&fq="+encode(fq)+"&sort="+sort);
+    q = SolrQuerySupport.toQuery("q="+encode(qs)+"&fq="+encode(fq)+"&sort="+sort);
     assertEquals(qs, q.getQuery());
-    assertEquals(new Integer(SolrRDD.DEFAULT_PAGE_SIZE), q.getRows());
+    assertEquals(new Integer(DEFAULT_PAGE_SIZE), q.getRows());
     assertTrue(q.getFilterQueries().length == 1);
     assertEquals(fq, q.getFilterQueries()[0]);
     List<SolrQuery.SortClause> sorts = q.getSorts();
