+++ src/test/java/com/lucidworks/spark/query/ShardSplitStrategyTest.java
@@ -1,8 +1,9 @@
 package com.lucidworks.spark.query;
 
 import com.lucidworks.spark.RDDProcessorTestBase;
-import com.lucidworks.spark.SolrRDD;
-import com.lucidworks.spark.SolrSupport;
+import com.lucidworks.spark.rdd.SolrRDD;
+import com.lucidworks.spark.util.SolrQuerySupport;
+import com.lucidworks.spark.util.SolrSupport;
 import org.apache.log4j.Logger;
 import org.apache.solr.client.solrj.SolrQuery;
 import org.apache.solr.client.solrj.response.QueryResponse;
@@ -81,8 +82,8 @@ public class ShardSplitStrategyTest extends RDDProcessorTestBase {
     assertTrue(minFromSort.longValue() == minFromStats.longValue());
     assertTrue(maxFromSort.longValue() == maxFromStats.longValue());
 
-    SolrRDD solrRDD = new SolrRDD(zkHost, collection);
-    List<String> shardList = solrRDD.buildShardList(cloudSolrServer);
+    SolrRDD solrRDD = new SolrRDD(zkHost, collection, jsc.sc());
+    List<String> shardList = SolrSupport.buildShardList(cloudSolrServer, collection);
 
     SolrQuery solrQuery = new SolrQuery("*:*");
     solrQuery.addFilterQuery("id:[* TO *]");
@@ -140,7 +141,7 @@ public class ShardSplitStrategyTest extends RDDProcessorTestBase {
         }
       }
 
-      QueryResponse qr = solrRDD.querySolr(SolrSupport.getHttpSolrClient(ss.getShardUrl()), splitQuery, 0, null);
+      QueryResponse qr = SolrQuerySupport.querySolr(SolrSupport.getHttpSolrClient(ss.getShardUrl()), splitQuery, 0, null);
       SolrDocumentList docList = qr.getResults();
       numDocs += docList.getNumFound();
       for (SolrDocument doc : docList) {
