+++ src/test/java/com/lucidworks/spark/query/StreamingResultsIteratorTest.java
@@ -1,13 +1,14 @@
 package com.lucidworks.spark.query;
 
 import com.lucidworks.spark.RDDProcessorTestBase;
-import com.lucidworks.spark.SolrRDD;
-import com.lucidworks.spark.SolrRDD.QueryResultsIterator;
+import com.lucidworks.spark.util.SolrQuerySupport;
+import com.lucidworks.spark.util.SolrSchemaUtil;
 import org.apache.solr.client.solrj.SolrQuery;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.impl.CloudSolrClient;
 import org.apache.solr.common.SolrDocument;
 import org.apache.solr.common.SolrInputDocument;
+import org.apache.spark.sql.types.StructType;
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -18,6 +19,7 @@ import java.util.List;
 import java.util.Random;
 
 import static org.junit.Assert.*;
+import static com.lucidworks.spark.util.SolrQuerySupport.*;
 
 public class StreamingResultsIteratorTest extends RDDProcessorTestBase {
 
@@ -73,8 +75,10 @@ public class StreamingResultsIteratorTest extends RDDProcessorTestBase {
     sendDocsThread.start();
     Thread.sleep(2000);
 
+    String uniqueKey = SolrQuerySupport.getUniqueKey(zkHost, testCollection);
+    StructType schema = SolrSchemaUtil.getBaseSchema(zkHost, testCollection, false);
     //StreamingResultsIterator sri = new StreamingResultsIterator(cloudSolrClient, solrQuery, "*");
-    QueryResultsIterator sri = new QueryResultsIterator(cloudSolrClient, solrQuery, "*");
+    QueryResultsIterator sri = new QueryResultsIterator(cloudSolrClient, solrQuery, "*") ;
     int numDocsFound = 0;
     boolean hasNext = false;
     do {
