+++ src/test/resources/conf/solrconfig.xml
@@ -4,7 +4,11 @@
   <dataDir>${solr.data.dir:}</dataDir>
   <directoryFactory name="DirectoryFactory" class="${solr.directoryFactory:solr.NRTCachingDirectoryFactory}"/>
   <codecFactory class="solr.SchemaCodecFactory"/>
-  <schemaFactory class="ClassicIndexSchemaFactory"/>
+<!--  <schemaFactory class="ClassicIndexSchemaFactory"/> -->
+  <schemaFactory class="ManagedIndexSchemaFactory">
+    <bool name="mutable">true</bool>
+    <str name="managedSchemaResourceName">managed-schema</str>
+  </schemaFactory>
   <indexConfig>
     <lockType>${solr.lock.type:native}</lockType>
     <unlockOnStartup>true</unlockOnStartup>
